refactor: added utils module, updated: imports
